+++ packages/desktop/views/login/LoginRouter.svelte
@@ -3,7 +3,7 @@
     import { LoginRoute, loginRoute } from '@core/router'
     import { UpdateStrongholdRouterView } from '@views'
     import { Transition } from '@ui'
-    import { EnterPinView, LoadProfileView, SelectProfileView } from './views'
+    import { LoginView, LoadProfileView, SelectProfileView } from './views'
     import features from '@features/features'
 
     $: if (features.analytics.loginRoute.enabled && $loginRoute)
@@ -16,7 +16,7 @@
     </Transition>
 {:else if $loginRoute === LoginRoute.EnterPin}
     <Transition>
-        <EnterPinView />
+        <LoginView />
     </Transition>
 {:else if $loginRoute === LoginRoute.UpdateStronghold}
     <Transition>
