+++ packages/desktop/views/login/views/index.js
@@ -1,3 +1,3 @@
-export { default as EnterPinView } from './EnterPinView.svelte'
+export { default as LoginView } from './LoginView.svelte'
 export { default as LoadProfileView } from './LoadProfileView.svelte'
 export { default as SelectProfileView } from './SelectProfileView.svelte'
